llcppsymg:pkg syspath -> symg/syspath.go
