--- _xtool/llcppsymg/_cmptest/syspath_test/llgo.expect
@@ -1,9 +1,0 @@
-#stdout
-=== TestLdOutput ===
-[/usr/local/lib/aarch64-linux-gnu /lib/aarch64-linux-gnu /usr/lib/aarch64-linux-gnu /usr/local/lib /lib /usr/lib /usr/aarch64-linux-gnu/lib]
-=== TestClangIncOutput ===
-[/usr/lib/llvm-18/lib/clang/18/include /usr/local/include /usr/include/aarch64-linux-gnu /usr/include]
-
-#stderr
-
-#exit 0
